Fixed bug related to pointer events not firing on custom labels correctly. Added test coverage for this functionality. Restructured test folder to prepare for increased testing coverage.
