+++ playwright.config.ts
@@ -5,7 +5,6 @@ const config: PlaywrightTestConfig = {
 		command: 'npm run build && npm run preview',
 		port: 4173
 	},
-
 	testDir: 'tests'
 };
